requirements.txt
mss>=9.0.1
pynput>=1.7.6
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0
openai>=1.30.0
python-dotenv>=1.0.0
pytest>=8.0.0
